+++ changelog.md
@@ -82,4 +82,9 @@ Changelog
 -	new command HGETKEYS, HGETVALS, HLEN
 -	optimized overwrite in lists
 
+### 1.3.0
+
+-	refactor pair to support values up to 64 GB
+-	refactor BITSET / BITGET
+-	new commands BITMSET, BITMGET